Change tab logic
