+++ src/components/TabList/TabList.jsx
@@ -1,19 +1,8 @@
-'use client';
-
-import { useState } from 'react';
 import PropTypes from 'prop-types';
 
 import { TabItem } from '../TabItem';
 
 export const TabList = ({ data, isBenefit = false, className }) => {
-  const [open, setOpen] = useState(null);
-
-  const toggle = index => {
-    if (open === index) {
-      return setOpen(null);
-    }
-    setOpen(index);
-  };
   return (
     <ul
       className={`${
@@ -21,19 +10,13 @@ export const TabList = ({ data, isBenefit = false, className }) => {
       } ${className}`}
     >
       {data.map((data, index) => (
-        <TabItem
-          key={index}
-          open={index === open}
-          isBenefit={isBenefit}
-          data={data}
-          toggle={() => toggle(index)}
-        />
+        <TabItem key={index} isBenefit={isBenefit} data={data} />
       ))}
     </ul>
   );
 };
 
-TabItem.propTypes = {
+TabList.propTypes = {
   isBenefit: PropTypes.bool,
   data: PropTypes.arrayOf(
     PropTypes.shape({
