+++ src/sections/Benefits/Benefits.jsx
@@ -3,7 +3,6 @@ import Image from 'next/image';
 import { Section } from '@/components/Section';
 import { TabList } from '@/components/TabList';
 import benefits from '@/data/benefits';
-import { TabListNew } from '@/components/TabList/TabListNew';
 
 export const Benefits = () => {
   return (
@@ -12,7 +11,7 @@ export const Benefits = () => {
         <div className="mt-[140px] notXl:hidden">
           <Image src="/smile.svg" width={452} height={311} alt="Smile Bar" />
         </div>
-        <TabListNew data={benefits} isBenefit />
+        <TabList data={benefits} isBenefit />
       </div>
     </Section>
   );
